refactor: remove commands folder
