+++ packages/core/src/lib/collect.ts
@@ -1,12 +1,10 @@
-import { CoreConfig, Report } from '@code-pushup/models';
+import { CoreConfig, GlobalOptions, Report } from '@code-pushup/models';
 import { calcDuration } from '@code-pushup/utils';
-import { name, version } from '../../../package.json';
-import { executePlugins } from '../implementation/execute-plugin';
+import { name, version } from '../../package.json';
+import { executePlugins } from './implementation/execute-plugin';
 
-export type CollectOptions = Pick<
-  CoreConfig,
-  'plugins' | 'categories' | 'upload'
->;
+export type CollectOptions = Pick<GlobalOptions, 'verbose'> &
+  Pick<CoreConfig, 'plugins' | 'categories' | 'upload'>;
 
 /**
  * Run audits, collect plugin output and aggregate it into a JSON object
